dbc2val/requirements-dev.txt
pyserial
cantools
python-can
pyyaml
j1939
py_expression_eval
grpcio
protobuf
types-PyYAML
types-protobuf
grpcio-tools
